backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
async-timeout==4.0.3
attrs==26.1.0
bcrypt==5.0.0
build==1.4.4
certifi==2026.4.22
cffi==2.0.0
charset-normalizer==3.4.7
chromadb==1.5.8
click==8.1.8
coloredlogs==15.0.1
cryptography==47.0.0
duckduckgo-search==8.1.1
durationpy==0.10
exceptiongroup==1.3.1
fastapi==0.128.8
filelock==3.19.1
flatbuffers==25.12.19
fsspec==2025.10.0
google-ai-generativelanguage==0.6.15
google-api-core==2.30.3
google-api-python-client==2.194.0
google-auth==2.49.2
google-auth-httplib2==0.3.1
google-generativeai==0.8.6
googleapis-common-protos==1.74.0
grpcio==1.80.0
grpcio-status==1.71.2
h11==0.16.0
hf-xet==1.4.3
httpcore==1.0.9
httplib2==0.31.2
httptools==0.7.1
httpx==0.28.1
huggingface_hub==1.8.0
humanfriendly==10.0
idna==3.13
importlib_metadata==8.7.1
importlib_resources==6.5.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kubernetes==35.0.0
langchain==0.3.28
langchain-core==0.3.84
langchain-text-splitters==0.3.11
langsmith==0.4.37
markdown-it-py==3.0.0
mdurl==0.1.2
mmh3==5.2.0
mpmath==1.3.0
numpy==2.0.2
oauthlib==3.3.1
onnxruntime==1.19.2
opentelemetry-api==1.41.1
opentelemetry-exporter-otlp-proto-common==1.41.1
opentelemetry-exporter-otlp-proto-grpc==1.41.1
opentelemetry-proto==1.41.1
opentelemetry-sdk==1.41.1
opentelemetry-semantic-conventions==0.62b1
orjson==3.11.5
overrides==7.7.0
packaging==25.0
proto-plus==1.27.2
protobuf==5.29.6
pyasn1==0.6.3
pyasn1_modules==0.4.2
pybase64==1.4.3
pycparser==2.23
pydantic==2.13.3
pydantic-settings==2.11.0
pydantic_core==2.46.3
Pygments==2.20.0
pyparsing==3.3.2
PyPika==0.51.1
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
PyYAML==6.0.3
referencing==0.36.2
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==15.0.0
rpds-py==0.27.1
shellingham==1.5.4
six==1.17.0
SQLAlchemy==2.0.49
starlette==0.49.3
sympy==1.14.0
tenacity==9.1.2
tokenizers==0.22.2
tomli==2.4.1
tqdm==4.67.3
typer==0.23.2
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
uuid_utils==0.14.1
uvicorn==0.39.0
uvloop==0.22.1
watchfiles==1.1.1
websocket-client==1.9.0
websockets==15.0.1
zipp==3.23.1
zstandard==0.25.0
